Return HTML data

+++ maps/models.py
@@ -6,7 +6,7 @@ class Country(models.Model):
     """Create Country model for Maps app."""
 
     name = models.CharField(max_length=200)
-    population = models.DecimalField(blank=True, null=True)
+    population = models.BigIntegerField(blank=True, null=True)
 
     def __str__(self):
         """Return country name in admin."""
